add registeration status modal
